requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.36.0
peft>=0.5.0
bitsandbytes>=0.41.0
accelerate>=0.21.0
optimum>=1.12.0
trl>=0.7.1
einops>=0.6.0
sentencepiece>=0.1.99
protobuf>=3.20.0
#flash-attn>=2.3.0  # Flash Attention for faster training
#triton>=2.0.0      # Required for Flash Attention

# Quantization libraries
bitsandbytes>=0.41.0

# Data processing
datasets>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Cloud storage and AWS
boto3>=1.28.0
awscli>=1.29.0

# Experiment tracking
mlflow>=2.8.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
wandb>=0.15.0

# Cloud and remote execution
paramiko>=3.3.0
requests>=2.31.0

# Development tools
black>=23.3.0
isort>=5.12.0
pytest>=7.3.1
numpy==1.24.3
